models/requirements.txt
# YOLOv8 Acne Detection Model Requirements
# Core ML libraries
tensorflow>=2.13.0
keras-cv>=0.6.0
numpy>=1.21.0
opencv-python>=4.8.0
Pillow>=9.0.0

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data processing
pandas>=1.5.0
matplotlib>=3.6.0
tqdm>=4.64.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0